Fix not found storm consumer groups when zookeeper-path=/
Burrow wasn't handling well the situation where storm consumer groups
were stored at the root of zookeeper `zookeeper-path=/`. It was trying
to get consumer group metadata building zookeeper path like this:
`//consumerGroupName` which was an invalid path. The fix handle this
situation by building the a path like: `/consumerGroupName` which is
valid.