Allow more generic HTTP client configuration
Allow the user to specify a different URL and HTTP method to close
alerts. Renamed POST/DELETE to be open/close as the HTTP method is now
configurable. The configuration file variable names still use `delete`
suffix instead of `close` to stay backwards compatible with the original
config file format.