Burrow as a library & minor refactoring
The state Burrow was in, I couldn't use it directly as a library. I've
seen a lot of projects using burrow as a separate binary and read from
another application. This is less than optimal if you need the Lag
evaluation bits in an existing Go monitoring tool.

This keeps the functionality exactly the same, but moves the files
around so that it will still continue to work as expected.

The package names have been changed to mine to use it with projects.