Optimization to #171
By always setting MaxOffset to the value of Offset when the ring entry is created, we assure that MaxOffset is always greater than or equal to Offset. This
means that when checking for whether or not we should insert an artificial commit, we only need one check - if MaxOffset is greater than or equal to the broker
offset.