Fix alert closing (#334)
This commit fixes two issues with alerts closed notifications not being sent.

The first is to only add a consumer group to the cluster map if it doesn't already exist.
If it already exists don't do anything. Previoulsy the groups were added fresh every time.
When they get added the startTime is set to 0. This prevents the closed logic from getting triggered

The second is a fix so that the closed template is used for closed alerts instead of the open template