Fix reporting final offset in backfill
Much related to #616 but for the actuall `__consumer_offsets` message.

This is not an issue if the consumer group is busy and committing often,
however if that is not the case it can lead to `Burrow` reporting
(a single consumer group topic partition) as being with status `STOP`
erroneously. This behaviour persists across restarts if no new messages
have been produced on the given `__consumer_offsets` partition.

\cc @bai & @timbertson :smile: