Correctly handle deletion of metrics when consumer group topic is deleted
Previously, we deleted the all consumer group metrics when a group's topic was deleted. WIth this change we make sure to check for the case where a topic is provided (i.e. user wants to delete only a specific topic of a consumer group) and delete only the metrics that match both consumer group and topic.